Rename view and serializer versions
