+++ EquiTrack/partners/urls.py
@@ -6,7 +6,7 @@ from django.conf.urls import patterns, url, include
 
 from rest_framework_nested import routers
 
-from .views.views import (
+from .views.v1 import (
     InterventionLocationView,
     PortalDashView,
     PortalLoginFailedView,
