+++ EquiTrack/partners/urls_v2.py
@@ -1,6 +1,6 @@
 from django.conf.urls import patterns, url
 
-from .views.views_v2 import (
+from .views.v2 import (
     AgreementListAPIView,
     AgreementDetailAPIView,
     AgreementInterventionsListAPIView,
